pyproject.toml
[tool.poetry]
name = "datasyn"
version = "0.1.0"
description = "Simple Data Ingestion from pub/sub -> bigquery"
authors = ["Antonio <antonio.one@pm.me>"]
license = "TBC"

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
google-cloud-pubsub = "^2.3.0"
python-decouple = "^3.4"
google-cloud-bigquery = "^2.8.0"
numpy = "^1.20.1"
pydantic = "^1.7.3"
click = "^7.1.2"
fastapi = {extras = ["all"], version = "^0.63.0"}

[tool.poetry.dev-dependencies]
pre-commit = "^2.10.0"
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
pytest-env = "^0.6.2"
black = "^20.8b1"
isort = "^5.7.0"
flake8 = "^3.8.4"
mypy = "^0.812"
PyYAML = "^5.4.1"
vcrpy = "^4.1.1"

[tool.isort]
profile = "black"

[tool.poetry.scripts]
synthesize = "datasyn.service_layer.publisher:publish"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
